+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ CLI changes
 
 Added
 ~~~~~
+* GeoJSON geometries consisting of multiple features are now supported. (`#530 <https://github.com/sentinelsat/sentinelsat/issues/530>`_ `@mackland <https://github.com/mackland>`_)
 * Better handling 403 error in ``trigger_offline_retrieval()`` (`#491 <https://github.com/sentinelsat/sentinelsat/issues/491>`_ `@z4zz <https://github.com/z4zz>`_)
 * Added support for SHA3-256 checksums. (`#523 <https://github.com/sentinelsat/sentinelsat/issues/523>`_ `@valgur <https://github.com/valgur>`_)